pyproject.toml
[tool.hatch.metadata]
allow-direct-references = true

[project]
name = "misaki"
version = "0.9.0"
description = "G2P engine for TTS"
readme = "README.md"
authors = [
    { name = "hexgrad", email = "hello@hexgrad.com" }
]
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8, <3.13"
dependencies = [
    "addict",
    "regex",
]

[project.optional-dependencies]
en = ["num2words", "spacy", "spacy-curated-transformers", "phonemizer-fork", "espeakng-loader"]
ja = ["fugashi", "jaconv", "mojimoji", "unidic-lite", "pyopenjtalk-plus"]
ko = ["jamo", "nltk"]
zh = ["jieba", "ordered-set", "pypinyin", "cn2an", "pypinyin-dict"]
vi = ["num2words", "spacy", "spacy-curated-transformers", "underthesea"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
